requirements.txt
fastapi[standard]
pydantic
sqlmodel
psycopg2-binary
python-dotenv
python-jose